debug generate

+++ call_scripts/train/train_1-run.sh
@@ -490,12 +490,17 @@ function pair_experiment() {
                                     --criterion nat_ctc_loss \
                                     --has-eos --max-update 70000 \
                                     --hydra \
-                                    -g 2 --fp16       
+                                    -g 2 --fp16   
 
     for experiment in $2 $3 $4; do
-        mkdir checkpoints/$experiment/
-        cp checkpoints/$1/checkpoint.best_bleu_* checkpoints/$experiment/
-        cp checkpoints/$1/checkpoint_last.pt checkpoints/$experiment/
+        if [ -e checkpoints/$experiment/checkpoint_last.pt ] && \
+        [ $(ls checkpoints/$experiment/checkpoint.best_bleu_* 2>/dev/null | grep -c "^checkpoints/$experiment/checkpoint.best_bleu_.*") -eq 5 ]; then    
+            echo "All 6 checkpoint files exist"
+        else 
+            mkdir checkpoints/$experiment/
+            cp checkpoints/$1/checkpoint.best_bleu_* checkpoints/$experiment/
+            cp checkpoints/$1/checkpoint_last.pt checkpoints/$experiment/     
+        fi     
     done
     
     for experiment in $1 $2 $3 $4; do
@@ -512,8 +517,16 @@ function pair_experiment() {
 
 }
 
-pair_experiment 2-2-1-1-H12-UF20M 2-2-1-1-N-UF20M 2-2-1-1-H7-UF20M 2-2-1-1-H4-UF20M
-pair_experiment 2-2-1-1-H12-UF20T 2-2-1-1-N-UF20T 2-2-1-1-H10-UF20T 2-2-1-1-H8-UF20M
+# pair_experiment 2-2-1-1-H12-UF20M 2-2-1-1-N-UF20M 2-2-1-1-H7-UF20M 2-2-1-1-H4-UF20M
+
+bash call_scripts/train_nat.sh -e 2-2-3-1-N-UR20M-rate_select-divTGT-NEW-3 \
+                                --save-interval-updates 70000 --max-tokens 1536 \
+                                --arch ctcpmlm_rate_selection \
+                                --task translation_ctcpmlm \
+                                --criterion nat_ctc_sel_rate_loss \
+                                --has-eos --max-update 100000 \
+                                --hydra \
+                                -g 2 --fp16
